Remove books manager at long last
