+++ bookwyrm/connectors/abstract_connector.py
@@ -1,24 +1,18 @@
 ''' functionality outline for a book data connector '''
 from abc import ABC, abstractmethod
 from dataclasses import asdict, dataclass
-import importlib
 import logging
 from urllib3.exceptions import RequestError
 
 from django.db import transaction
 import requests
-from requests import HTTPError
 from requests.exceptions import SSLError
 
 from bookwyrm import activitypub, models, settings
-from bookwyrm.tasks import app
+from .connector_manager import load_more_data, ConnectorException
 
 
 logger = logging.getLogger(__name__)
-class ConnectorException(HTTPError):
-    ''' when the connector can't do what was asked '''
-
-
 class AbstractMinimalConnector(ABC):
     ''' just the bare bones, for other bookwyrm instances '''
     def __init__(self, identifier):
@@ -190,23 +184,6 @@ class AbstractConnector(AbstractMinimalConnector):
     @abstractmethod
     def expand_book_data(self, book):
         ''' get more info on a book '''
-
-
-@app.task
-def load_more_data(connector_id, book_id):
-    ''' background the work of getting all 10,000 editions of LoTR '''
-    connector_info = models.Connector.objects.get(id=connector_id)
-    connector = load_connector(connector_info)
-    book = models.Book.objects.select_subclasses().get(id=book_id)
-    connector.expand_book_data(book)
-
-
-def load_connector(connector_info):
-    ''' instantiate the connector class '''
-    connector = importlib.import_module(
-        'bookwyrm.connectors.%s' % connector_info.connector_file
-    )
-    return connector.Connector(connector_info.identifier)
 
 
 def dict_from_mappings(data, mappings):
