+++ bookwyrm/tests/models/test_import_model.py
@@ -8,7 +8,8 @@ from django.utils import timezone
 from django.test import TestCase
 import responses
 
-from bookwyrm import books_manager, models
+from bookwyrm import models
+from bookwyrm.connectors import connector_manager
 from bookwyrm.connectors.abstract_connector import SearchResult
 
 
@@ -134,7 +135,7 @@ class ImportJob(TestCase):
             search_url='https://openlibrary.org/search?q=',
             priority=3,
         )
-        connector = books_manager.load_connector(connector_info)
+        connector = connector_manager.load_connector(connector_info)
         result = SearchResult(
             title='Test Result',
             key='https://openlibrary.org/works/OL1234W',
@@ -163,7 +164,7 @@ class ImportJob(TestCase):
             json={'name': 'test author'},
             status=200)
 
-        with patch('bookwyrm.books_manager.first_search_result') as search:
+        with patch('bookwyrm.connector_manager.first_search_result') as search:
             search.return_value = result
             book = self.item_1.get_book_from_isbn()
